Rename utils.cog module to be consistent with other utils
