+++ cogs/Admin/admin.py
@@ -8,7 +8,7 @@ import re
 import discord
 from discord.ext import commands
 
-from ..utils.cog import BasicCog
+from ..utils.cogs import BasicCog
 
 
 logger = logging.getLogger('discord')
