+++ cogs/Roles/roles.py
@@ -4,7 +4,7 @@ from discord.ext import commands
 
 import numpy as np
 
-from ..utils.cog import BasicCog
+from ..utils.cogs import BasicCog
 
 
 class Roles(BasicCog):
